refactor: extract components from Your Information
