--- components/account/EditableContactSection.tsx
@@ -1,1 +1,0 @@
-// TODO: Create the Editable Contact Section component from the Your Information page
