--- components/account/PersonFieldsSection.tsx
@@ -1,1 +1,0 @@
-// TODO: Create the Person Fields Section component from the Your Information page
