+++ pages/account/welcome/your-information.tsx
@@ -13,27 +13,22 @@ import { useRouter } from "next/router";
 import { Trans, useTranslation } from "next-i18next";
 import { serverSideTranslations } from "next-i18next/serverSideTranslations";
 
-import { FC, useContext, useState } from "react";
+import { useContext, useState } from "react";
 
 // SK Components
 import {
   Actions,
   Button,
-  Card,
-  Columns,
   ContentLayout,
   Header,
   MaterialIcon,
-  MenuItem,
   Section,
-  Select,
   Snackbar,
-  TextField,
 } from "@suankularb-components/react";
 
 // Internal components
-import ContactCard from "@/components/account/ContactCard";
-import ContactDialog from "@/components/account/ContactDialog";
+import ContactsSection from "@/components/account/ContactSection";
+import PersonFields from "@/components/account/PersonFields";
 import MySKPageHeader from "@/components/common/MySKPageHeader";
 import NextWarningCard from "@/components/welcome/NextWarningCard";
 
@@ -41,271 +36,23 @@ import NextWarningCard from "@/components/welcome/NextWarningCard";
 import SnackbarContext from "@/contexts/SnackbarContext";
 
 // Backend
-import { getPersonFromUser, setupPerson } from "@/utils/backend/person/person";
+import { getPersonFromUser, editPerson } from "@/utils/backend/person/person";
 import { getSubjectGroups } from "@/utils/backend/subject/subjectGroup";
 
 // Helpers
 import { changeItem } from "@/utils/helpers/array";
-import { getLocaleString } from "@/utils/helpers/i18n";
 import { withLoading } from "@/utils/helpers/loading";
 import { createTitleStr } from "@/utils/helpers/title";
 
 // Hooks
 import { useForm } from "@/utils/hooks/form";
-import { useLocale } from "@/utils/hooks/i18n";
 import { useToggle } from "@/utils/hooks/toggle";
 
 // Types
-import { CustomPage, FormControlProps, LangCode } from "@/utils/types/common";
+import { CustomPage, LangCode } from "@/utils/types/common";
 import { Contact } from "@/utils/types/contact";
 import { Student, Teacher } from "@/utils/types/person";
 import { SubjectGroup } from "@/utils/types/subject";
-
-const ThaiNameSection: FC<{ formProps: FormControlProps }> = ({
-  formProps,
-}) => {
-  // Translation
-  const { t } = useTranslation("account");
-
-  return (
-    <Section>
-      <Header level={3}>{t("profile.name.title")}</Header>
-      <Columns columns={4} className="my-3 !gap-y-12">
-        <TextField
-          appearance="outlined"
-          label={t("profile.name.prefix")}
-          helperMsg={t("profile.name.prefix_helper")}
-          {...formProps.prefixTH}
-        />
-        <TextField
-          appearance="outlined"
-          label={t("profile.name.firstName")}
-          {...formProps.firstNameTH}
-        />
-        <TextField
-          appearance="outlined"
-          label={t("profile.name.middleName")}
-          helperMsg={t("profile.name.middleName_helper")}
-          {...formProps.middleNameTH}
-        />
-        <TextField
-          appearance="outlined"
-          label={t("profile.name.lastName")}
-          {...formProps.lastNameTH}
-        />
-        {/* <TextField
-          appearance="outlined"
-          label={t("profile.name.nickname")}
-          {...formProps.nicknameTH}
-        /> */}
-      </Columns>
-    </Section>
-  );
-};
-
-const EnglishNameSection: FC<{ formProps: FormControlProps }> = ({
-  formProps,
-}) => {
-  // Translation
-  const { t } = useTranslation("account");
-
-  return (
-    <Section>
-      <Header level={3}>{t("profile.enName.title")}</Header>
-      <Columns columns={4} className="my-3 !gap-y-12">
-        <TextField
-          appearance="outlined"
-          label={t("profile.enName.prefix")}
-          helperMsg={t("profile.enName.prefix_helper")}
-          {...formProps.prefixEN}
-        />
-        <TextField
-          appearance="outlined"
-          label={t("profile.enName.firstName")}
-          {...formProps.firstNameEN}
-        />
-        <TextField
-          appearance="outlined"
-          label={t("profile.enName.middleName")}
-          helperMsg={t("profile.name.middleName_helper")}
-          {...formProps.middleNameEN}
-        />
-        <TextField
-          appearance="outlined"
-          label={t("profile.enName.lastName")}
-          {...formProps.lastNameEN}
-        />
-        {/* <TextField
-          appearance="outlined"
-          label={t("profile.enName.nickname")}
-          {...formProps.nicknameEN}
-        /> */}
-      </Columns>
-    </Section>
-  );
-};
-
-const RoleSection: FC<{
-  formProps: FormControlProps;
-  subjectGroups: SubjectGroup[];
-}> = ({ formProps, subjectGroups }) => {
-  // Translation
-  const locale = useLocale();
-  const { t } = useTranslation("account");
-
-  return (
-    <Section>
-      <Header level={3}>{t("profile.role.title")}</Header>
-      <Columns columns={4} className="mb-8 mt-3 !gap-y-12">
-        <Select
-          appearance="outlined"
-          label={t("profile.role.subjectGroup")}
-          helperMsg={t("profile.role.subjectGroup_helper")}
-          {...formProps.subjectGroup}
-        >
-          {subjectGroups.map((subjectGroup) => (
-            <MenuItem key={subjectGroup.id} value={subjectGroup.id}>
-              {getLocaleString(subjectGroup.name, locale)}
-            </MenuItem>
-          ))}
-        </Select>
-        <TextField
-          appearance="outlined"
-          label={t("profile.role.classAdvisorAt")}
-          helperMsg={t("profile.role.classAdvisorAt_helper")}
-          {...formProps.classAdvisorAt}
-        />
-      </Columns>
-    </Section>
-  );
-};
-
-const MiscellaneousSection: FC<{ formProps: FormControlProps }> = ({
-  formProps,
-}) => {
-  // Translation
-  const { t } = useTranslation("account");
-
-  return (
-    <Section>
-      <Header level={3}>{t("profile.general.title")}</Header>
-      <Columns columns={4} className="my-3 !gap-y-12">
-        {/* <Select
-          appearance="outlined"
-          label={t("profile.general.gender.label")}
-          {...formProps.gender}
-        >
-          <MenuItem value="male">{t("profile.general.gender.male")}</MenuItem>
-          <MenuItem value="female">
-            {t("profile.general.gender.female")}
-          </MenuItem>
-          <MenuItem value="non-binary">
-            {t("profile.general.gender.nonBinary")}
-          </MenuItem>
-          <MenuItem value="no-response">
-            {t("profile.general.gender.noReponse")}
-          </MenuItem>
-        </Select> */}
-        <TextField
-          appearance="outlined"
-          label={t("profile.general.birthDate")}
-          inputAttr={{ type: "date" }}
-          {...formProps.birthdate}
-        />
-        {/* <TextField
-          appearance="outlined"
-          label={t("profile.general.citizenID")}
-          leading={<MaterialIcon icon="lock" />}
-          helperMsg={t("profile.common.privateInfo_helper")}
-          inputAttr={{ type: "number" }}
-          {...formProps.citizenID}
-        /> */}
-        {/* <TextField
-          appearance="outlined"
-          label={t("profile.general.passportNumber")}
-          leading={<MaterialIcon icon="lock" />}
-          helperMsg={t("profile.common.privateInfo_helper")}
-          {...formProps.passportNumber}
-        /> */}
-        {/* <Select
-          appearance="outlined"
-          label={t("profile.general.bloodGroup")}
-          leading={<MaterialIcon icon="lock" />}
-          helperMsg={t("profile.common.privateInfo_helper")}
-          {...formProps.bloodGroup}
-        >
-          <MenuItem value="A">A</MenuItem>
-          <MenuItem value="B">B</MenuItem>
-          <MenuItem value="AB">AB</MenuItem>
-          <MenuItem value="O">O</MenuItem>
-        </Select> */}
-      </Columns>
-    </Section>
-  );
-};
-
-const ContactsSection: FC<{
-  contacts: Contact[];
-  setContacts: (value: Contact[]) => void;
-}> = ({ contacts, setContacts }) => {
-  // Translation
-  const { t } = useTranslation("account");
-
-  // Dialog control
-  const [showAdd, setShowAdd] = useState<boolean>(false);
-
-  return (
-    <Section sectionAttr={{ "aria-labelledby": "header-contacts" }}>
-      <Columns columns={3} className="!items-end">
-        <Header className="md:col-span-2" hAttr={{ id: "header-contacts" }}>
-          {t("profile.contacts.title")}
-        </Header>
-        <Actions>
-          <Button
-            appearance="tonal"
-            icon={<MaterialIcon icon="add" />}
-            onClick={() => setShowAdd(true)}
-          >
-            {t("profile.contacts.action.add")}
-          </Button>
-          <ContactDialog
-            open={showAdd}
-            onClose={() => setShowAdd(false)}
-            onSubmit={(contact) => {
-              setShowAdd(false);
-              setContacts([...contacts, contact]);
-            }}
-          />
-        </Actions>
-      </Columns>
-      {contacts.length ? (
-        <Columns columns={4}>
-          {contacts.map((contact, idx) => (
-            <ContactCard
-              key={contact.id}
-              contact={contact}
-              onChange={(contact) =>
-                setContacts(changeItem<Contact>(contact, idx, contacts))
-              }
-              onRemove={() =>
-                setContacts(contacts.filter((item) => contact.id !== item.id))
-              }
-            />
-          ))}
-        </Columns>
-      ) : (
-        <Card
-          appearance="outlined"
-          className="box-content !grid h-[4.5rem] place-content-center"
-        >
-          <p className="skc-body-medium text-on-surface-variant">
-            {t("profile.contacts.noContacts")}
-          </p>
-        </Card>
-      )}
-    </Section>
-  );
-};
 
 const WelcomePage: CustomPage<{
   person: Student | Teacher;
@@ -406,7 +153,7 @@ const WelcomePage: CustomPage<{
   const [loading, toggleLoading] = useToggle();
   async function handleSubmit() {
     withLoading(async () => {
-      const { error } = await setupPerson(
+      const { error } = await editPerson(
         supabase,
         { ...form, contacts },
         person
@@ -446,16 +193,18 @@ const WelcomePage: CustomPage<{
               </Link>
             </Trans>
           </p>
-          <ThaiNameSection formProps={formProps} />
-          <EnglishNameSection formProps={formProps} />
-          {person.role === "teacher" ? (
-            <RoleSection formProps={formProps} subjectGroups={subjectGroups} />
-          ) : (
-            <></>
-          )}
-          <MiscellaneousSection formProps={formProps} />
+          <PersonFields {...{ subjectGroups, formProps }} />
         </Section>
-        <ContactsSection contacts={contacts} setContacts={setContacts} />
+        <ContactsSection
+          contacts={contacts}
+          handleAdd={(contact) => setContacts([...contacts, contact])}
+          handleEdit={(contact, idx) =>
+            setContacts(changeItem<Contact>(contact, idx, contacts))
+          }
+          handleRemove={(contactID) =>
+            setContacts(contacts.filter((item) => contactID !== item.id))
+          }
+        />
         <Actions className="mx-4 sm:mx-0 sm:mb-20">
           <Button
             appearance="filled"
